✨ Add a github workflow to make build files
They are not currently being deployed anyhwere, and the build system
itself leaves a lot to be desired. (The original build script was much
better, honestly)

✨ Move to just using a Makefile since bikeshed works with Python 3 finally
🔥 Delete unwritten vfs file
🔥 Delete old build script
🐛 Fix syntax in integer-width-literals

TODO:
Some more work is needed to make sure we can
1) Output a numbered version of the paper
2) Provide a more accurate naming convention for accepted, rejected,
   delayed, withdrawn, active, draft, etc. This might just end up being a
   bunch of different directories, as rejected, accepted, withdrawn
   aren't going to be built anytime soon, and we can rely on github's
   file cacheing for its output.

Additionally, would it honestly just be easier to use CMake for this? I
feel like it would, but I'm not so sure